+++ src/db/migrate/20210907061744_create_recipefoods.rb
@@ -2,14 +2,14 @@ class CreateRecipefoods < ActiveRecord::Migration[5.2]
   def change
     create_table :recipefoods do |t|
       t.string :food_name
-      t.integer :amount
-      t.integer :calorie
-      t.integer :protein
-      t.integer :fat
-      t.integer :carbo
-      t.integer :suger
-      t.integer :dietary_fiber
-      t.integer :salt
+      t.float :amount
+      t.float :calorie
+      t.float :protein
+      t.float :fat
+      t.float :carbo
+      t.float :suger
+      t.float :dietary_fiber
+      t.float :salt
       t.references :user, foreign_key: true
       t.references :recipe
