+++ src/db/migrate/20210908142518_create_food_menus.rb
@@ -2,13 +2,13 @@ class CreateFoodMenus < ActiveRecord::Migration[5.2]
   def change
     create_table :food_menus do |t|
       t.string :food_name
-      t.integer :amount
-      t.integer :calorie
-      t.integer :protein
-      t.integer :fat
-      t.integer :carbo
-      t.integer :suger
-      t.integer :salt
+      t.float :amount
+      t.float :calorie
+      t.float :protein
+      t.float :fat
+      t.float :carbo
+      t.float :suger
+      t.float :salt
 
       t.timestamps
     end
