+++ src/db/migrate/20210908143325_create_my_meals.rb
@@ -3,13 +3,13 @@ class CreateMyMeals < ActiveRecord::Migration[5.2]
     create_table :my_meals do |t|
       t.datetime :start_time
       t.string :food_name
-      t.integer :calorie
+      t.float :calorie
       t.string :protein
       t.string :fat
       t.string :carbo
-      t.integer :suger
-      t.integer :dietary_fiber
-      t.integer :salt
+      t.float :suger
+      t.float :dietary_fiber
+      t.float :salt
       t.string :note
       t.references :timezone
       t.references :user, foreign_key: true
